requirements.txt
numpy
opencv-python
paddlepaddle-gpu
paddleocr
pandas
Pillow
argparse